Move language pack downloads out of postinstall into npm run make
